move NamingStrategy to inside Stategy
